AGENTE-IASS-VIKING/pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
flask = "^3.0.3"
openai = "^1.29.0"
packaging = "^24.0"
functions = "^0.7.0"
requests = "^2.32.3"
pytz = "^2024.1"
anthropic = "^0.30.0"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
sift-stack-py = "^0.3.1"
google-ai-generativelanguage = "^0.6.14"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"